tally-ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai==1.3.5
pandas==2.1.3
scikit-learn==1.3.2